# Upgrade to JDK 24, Netty 4.2 and Virtual Thread Experimental support