--- a/frameworks/Java/netty/.sdkmanrc
+++ b/frameworks/Java/netty/.sdkmanrc
@@ -1,0 +1,3 @@
+# Enable auto-env through the sdkman_auto_env config
+# Add key=value pairs of SDKs to use below
+java=24-oracle