--- a/frameworks/Java/netty/run_netty_loom.sh
+++ b/frameworks/Java/netty/run_netty_loom.sh
@@ -1,0 +1,15 @@
+#!/bin/bash
+
+# PROFILING: -XX:+UnlockDiagnosticVMOptions -XX:+DebugNonSafepoints
+JAVA_OPTIONS="--enable-native-access=ALL-UNNAMED \
+  -Dio.netty.noUnsafe=false \
+  --sun-misc-unsafe-memory-access=allow \
+  --add-opens=java.base/java.lang=ALL-UNNAMED \
+  -XX:+UseNUMA \
+  -XX:+UseParallelGC \
+  -Dio.netty.buffer.checkBounds=false \
+  -Dio.netty.buffer.checkAccessible=false \
+  -Dhello.eventloop.carrier=true \
+  $@"
+
+java $JAVA_OPTIONS -jar app.jar