--- a/frameworks/Java/netty/run_netty_loom.sh
+++ b/frameworks/Java/netty/run_netty_loom.sh
@@ -10,6 +10,9 @@ JAVA_OPTIONS="--enable-native-access=ALL-UNNAMED \
   -Dio.netty.buffer.checkBounds=false \
   -Dio.netty.buffer.checkAccessible=false \
   -Dhello.eventloop.carrier=true \
+  -XX:+UnlockExperimentalVMOptions \
+  -XX:-DoJVMTIVirtualThreadTransitions \
+  -Djdk.trackAllThreads=false \
   $@"
 
 java $JAVA_OPTIONS -jar app.jar