--- a/frameworks/Java/netty/src/main/java/hello/loom/HelloLoomServerInitializer.java
+++ b/frameworks/Java/netty/src/main/java/hello/loom/HelloLoomServerInitializer.java
@@ -1,0 +1,21 @@
+package hello.loom;
+
+import java.util.concurrent.ScheduledExecutorService;
+
+import hello.HelloServerHandler;
+import hello.HelloServerInitializer;
+
+public class HelloLoomServerInitializer extends HelloServerInitializer {
+
+   private final MultithreadVirtualEventExecutorGroup group;
+
+   public HelloLoomServerInitializer(MultithreadVirtualEventExecutorGroup group, ScheduledExecutorService service) {
+      super(service);
+      this.group = group;
+   }
+
+   @Override
+   protected HelloServerHandler newHelloServerHandler(ScheduledExecutorService service) {
+      return new VirtualThreadHelloServerHandler(service, group);
+   }
+}