--- a/frameworks/Java/netty/src/main/java/hello/loom/LoomSupport.java
+++ b/frameworks/Java/netty/src/main/java/hello/loom/LoomSupport.java
@@ -4,9 +4,6 @@ import java.lang.invoke.MethodHandle;
 import java.lang.invoke.MethodHandles;
 import java.lang.reflect.Field;
 import java.util.concurrent.Executor;
-import java.util.concurrent.ForkJoinPool;
-
-import io.netty.util.internal.PlatformDependent;
 
 public final class LoomSupport {
    private static final boolean SUPPORTED;