This is both upgrading to Netty 4.2, unifying few existing code paths and introducing a novel "feature" for Netty 
i.e. event loop as carriers for https://openjdk.org/jeps/444

The feature here is used in an highly opinionated way, ensuring that pipelining is still possible and effective.