@franz1981 2025-04-10
This is blocked by https://github.com/netty/netty/commit/620794f458549b0d228715ebf4b359fdabe46140