@franz1981 2025-04-11
Ok it seems https://github.com/netty/netty/commit/620794f458549b0d228715ebf4b359fdabe46140 is not as harmfull as expected and we have some decent flamegraphs for this both with/without Loom - although we need to put some extra care while configure it because:
- virtual thread containers tracking uses a concurrent hash map under the hood
- virtual thread JVMTI state management (while attaching a profiler/agent) can still harms, but only while profiling 

The current Loom test is disabling some of these features to avoid been bitten.