@franz1981 2025-04-11
FYI plaintext on my machine (with fewer connections/threads etc etc) reports
with event loop carrier
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   144.50us    2.09ms  78.53ms   82.03%
    Req/Sec     0.94M   127.60k    1.22M    81.17%
  Latency Distribution
     50%   59.00us
     75%   88.00us
     90%    0.00us
     99%    0.00us
  74938880 requests in 40.10s, 9.14GB read
Requests/sec: 1868835.89
Transfer/sec:    233.48MB
```
without event loop carrier:
```
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   150.18us    2.16ms  79.88ms   78.56%
    Req/Sec     1.12M   152.01k    1.34M    86.16%
  Latency Distribution
     50%   52.00us
     75%   90.00us
     90%    0.00us
     99%    0.00us
  88888048 requests in 40.10s, 10.84GB read
Requests/sec: 2216695.70
Transfer/sec:    276.93MB
```
which is indeed a clear signal of the overhead of allocating threads with pipelining (i.e. in tight loops)